+++ tests/Codeception/Acceptance/_bootstrap.php
@@ -8,9 +8,10 @@
 declare(strict_types=1);
 
 // This is acceptance bootstrap
+use OxidEsales\EshopCommunity\Internal\Framework\FileSystem\ProjectRootLocator;
 use Symfony\Component\Filesystem\Path;
 
-require_once Path::join((new \OxidEsales\Facts\Facts())->getShopRootPath(), 'source', 'bootstrap.php');
+require_once Path::join((new ProjectRootLocator())->getProjectRoot(), 'source', 'bootstrap.php');
 
 $helper = new \OxidEsales\Codeception\Module\FixturesHelper();
 $helper->loadRuntimeFixtures(__DIR__ . '/../Support/Data/fixtures.php');