+++ tests/Integration/Greeting/Repository/GreetingRepositoryTest.php
@@ -10,23 +10,31 @@ declare(strict_types=1);
 namespace OxidEsales\ModuleTemplate\Tests\Integration\Greeting\Repository;
 
 use OxidEsales\Eshop\Application\Model\User as EshopModelUser;
+use OxidEsales\EshopCommunity\Core\Di\ContainerFacade;
 use OxidEsales\EshopCommunity\Internal\Framework\Database\QueryBuilderFactoryInterface;
-use OxidEsales\EshopCommunity\Tests\Integration\IntegrationTestCase;
 use OxidEsales\ModuleTemplate\Greeting\Repository\GreetingRepository;
 use OxidEsales\ModuleTemplate\Greeting\Repository\GreetingRepositoryInterface;
+use PHPUnit\Framework\TestCase;
 use PHPUnit\Framework\Attributes\CoversClass;
 
 #[CoversClass(GreetingRepository::class)]
-class GreetingRepositoryTest extends IntegrationTestCase
+class GreetingRepositoryTest extends TestCase
 {
     public const TEST_USER_ID = '_testuser';
     public const TEST_GREETING = 'Hi there';
+
+    protected function tearDown(): void
+    {
+        $this->cleanUpUsers();
+
+        parent::tearDown();
+    }
 
     public function testGetSavedUserGreeting(): void
     {
         $this->prepareTestData();
 
-        $repo = $this->get(GreetingRepositoryInterface::class);
+        $repo = ContainerFacade::get(GreetingRepositoryInterface::class);
 
         $this->assertSame(self::TEST_GREETING, $repo->getSavedUserGreeting(self::TEST_USER_ID));
         $this->assertSame('', $repo->getSavedUserGreeting('_notexisting'));
@@ -34,13 +42,11 @@ class GreetingRepositoryTest extends IntegrationTestCase
 
     private function prepareTestData(): void
     {
-        $this->cleanUpUsers();
-
         $user = oxNew(EshopModelUser::class);
         $user->assign(
             [
-                'oxid'         => self::TEST_USER_ID,
-                'oemtgreeting' => self::TEST_GREETING,
+                'oxid' => self::TEST_USER_ID,
+                'oemtgreeting' => self::TEST_GREETING
             ]
         );
         $user->save();
@@ -48,7 +54,7 @@ class GreetingRepositoryTest extends IntegrationTestCase
 
     private function cleanUpUsers()
     {
-        $queryBuilder = $this->get(QueryBuilderFactoryInterface::class)->create();
+        $queryBuilder = ContainerFacade::get(QueryBuilderFactoryInterface::class)->create();
         $queryBuilder->delete('oxuser');
         $queryBuilder->execute();
     }