+++ tests/Integration/IntegrationTestCase.php
@@ -10,9 +10,10 @@ declare(strict_types=1);
 namespace OxidEsales\ModuleTemplate\Tests\Integration;
 
 use OxidEsales\EshopCommunity\Internal\Container\ContainerFactory;
+use Codeception\PHPUnit\TestCase;
 use OxidEsales\EshopCommunity\Internal\Framework\Database\QueryBuilderFactoryInterface;
 
-class IntegrationTestCase extends \OxidEsales\EshopCommunity\Tests\Integration\IntegrationTestCase
+class IntegrationTestCase extends TestCase
 {
     protected function cleanUpUsers()
     {
@@ -25,6 +26,13 @@ class IntegrationTestCase extends \OxidEsales\EshopCommunity\Tests\Integration\I
     {
         $queryBuilder = $this->get(QueryBuilderFactoryInterface::class)->create();
         $queryBuilder->delete('oemt_tracker');
+        $queryBuilder->execute();
+    }
+
+    protected function cleanUpVotes()
+    {
+        $queryBuilder = $this->get(QueryBuilderFactoryInterface::class)->create();
+        $queryBuilder->delete('oemt_product_vote');
         $queryBuilder->execute();
     }